making some changes

+++ functions/utils/utils_strlen.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   utils_strlen.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: abdael-m <abdael-m@student.42.fr>          +#+  +:+       +#+        */
+/*   By: bnafiai <bnafiai@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/15 22:23:40 by abdael-m          #+#    #+#             */
-/*   Updated: 2025/04/05 08:48:59 by abdael-m         ###   ########.fr       */
+/*   Updated: 2025/04/13 19:13:23 by bnafiai          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-int	utils_strlen(const char *string)
+size_t	utils_strlen(const char *string)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (string[count] != '\0')
